requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
pydantic-settings==2.2.1
PyJWT==2.8.0
supabase==2.3.1
supabase-auth==0.5.5
postgrest-py==0.11.5
pytz==2024.1
tzdata==2024.1
